Mask user details in Sentry exception logs
## What

This changes configures Raven, the client for Sentry, so that PII data is taken out of the error report when sent to Sentry.

Specifically the entry in the hash `[:sidekiq][:job][:args][:arguments]` and `[:sidekiq][:jobstr]` which include name and telephone number .

## Why

We don't want to store or transfer any data unecessarily.
Sidekiq automatically logs some arguments we use for the job