Preserve query params when redirecting from /start
Trello: https://trello.com/c/4LpwmmCy

What's changed?
---------------

Pass any and all query string parameters from `/start` to the first
question.

For example: `/start?foo=123` now redirects to
`/live-in-england?foo=123` rather than `/live-in-england`

If no query params are passed in, `/start` redirects to
`/medical-equipment` as before.

Why?
----

We need to preserve the `_ga` parameter from GOV.UK to the service for
cross domain tracking to work.
The original 301 redirect broke this. I think it's because
analytics.js.erb isn't being called until a page in the form is loaded,
by which point we'd already lost the query params with the `_ga`
parameter.

How to test locally
-------------------

Run the app:

`foreman start`

Go to `localhost:5000/start?foo=123`

You should be redirected to `localhost:5000/live-in-england?foo=123`