Validate selected address
It is possible to change the address chosen via the address selection
dropdown.  We should validate and remove any such changes before progessing
onto the next step.

This change removes unexpected fields and any fields that do not have string
values.  If any changes are detected, the UPRN is also removed as we can no
longer trust it.